requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --index-url=https://mirrors.aliyun.com/pypi/simple/ requirements.in
#
--index-url https://mirrors.aliyun.com/pypi/simple/

aiohttp==3.8.5
    # via openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via uvicorn
decorator==4.4.2
    # via moviepy
exceptiongroup==1.1.3
    # via anyio
fastapi==0.100.1
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
imageio==2.31.1
    # via moviepy
imageio-ffmpeg==0.4.8
    # via moviepy
moviepy==1.0.3
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.25.2
    # via
    #   -r requirements.in
    #   imageio
    #   moviepy
openai==0.27.8
    # via -r requirements.in
pillow==10.0.0
    # via
    #   -r requirements.in
    #   imageio
proglog==0.1.10
    # via moviepy
pydantic==2.2.0
    # via fastapi
pydantic-core==2.6.0
    # via pydantic
python-dotenv==1.0.0
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   moviepy
    #   openai
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.19
    # via -r requirements.in
starlette==0.27.0
    # via fastapi
tqdm==4.65.0
    # via
    #   -r requirements.in
    #   moviepy
    #   openai
    #   proglog
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
urllib3==2.0.4
    # via requests
uvicorn==0.23.2
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp
retrying==1.3.4
    # via -r requirements.in
six==1.16.0
    # via retrying
harvesttext==1.25.2
edge-tts==6.1.9
easyocr==1.7.0